Completely restructured app to make easier to work with
