+++ app/__init__.py
@@ -1,0 +1,33 @@
+import sentry_sdk
+
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+from flask_login import LoginManager
+from flask_bcrypt import Bcrypt
+from sentry_sdk.integrations.flask import FlaskIntegration
+
+from config import Config
+
+db = SQLAlchemy()
+migrate = Migrate()
+login = LoginManager()
+bcrypt = Bcrypt()
+
+
+def create_app(config_class=Config):
+    app = Flask(__name__, static_folder='static')
+    app.config.from_object(Config)
+
+    sentry_sdk.init(dsn=app.config["SENTRY_DSN"], integrations=[FlaskIntegration()])
+
+    db.init_app(app)
+    migrate.init_app(app, db)
+    login.init_app(app)
+    login.login_view = "main.login"
+    bcrypt.init_app(app)
+
+    from app.views import bp
+    app.register_blueprint(bp)
+
+    return app
