+++ app/templates/base.html
@@ -12,14 +12,14 @@
 <body class="bg-gray-100 text-gray-900 tracking-wider leading-normal">
 <nav class="flex items-center justify-between flex-wrap bg-teal-500 p-6">
     <div class="flex items-center flex-shrink-0 text-white mr-6">
-        <a href="{{ url_for("index") }}" class="font-semibold text-xl tracking-tight">URL Shortener</a>
+        <a href="{{ url_for("main.index") }}" class="font-semibold text-xl tracking-tight">URL Shortener</a>
     </div>
     <div class="w-full block flex-grow lg:flex lg:items-center lg:w-auto">
         <div class="text-sm lg:flex-grow">
         </div>
         <div>
             {% if current_user.is_authenticated %}
-                <a href="{{ url_for("logout") }}"
+                <a href="{{ url_for("main.logout") }}"
                    class="inline-block text-sm px-4 py-2 leading-none border rounded text-white border-white hover:border-transparent hover:text-teal-500 hover:bg-white mt-4 lg:mt-0">Logout</a>
             {% endif %}
         </div>
